Remove SizedIdProvider

+++ src/send.rs
@@ -37,7 +37,7 @@ where
         let mut request_count = 0;
         for request in requests {
             // route to correct thread; share the load based on id and the mod of the thread count
-            let targeted = request.id() as usize % thread_count;
+            let targeted = request.id() % thread_count;
             event!(Level::DEBUG, "Sending to target {}", request.id());
             event!(Level::TRACE, ?request);
             guard[targeted].send(&send_back_to.clone(), request);
